requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
Django==2.2.4
django-filter==2.2.0
djangorestframework==3.10.2
future==0.17.1
idna==2.8
Pillow==6.1.0
psycopg2-binary==2.8.3
pycparser==2.19
python-telegram-bot==12.0.0b1
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
tornado==6.0.3
urllib3==1.25.3
